Data-Driven Subsampling Based on Uniform Design

A model-free subsampling method based on uniform design proposed by Zhang et.al. (2023).

Usage

DDS(X, n, ratio = 0.85)

Arguments

X

A data.frame or matrix of explanatory variables.

n

Subsample size.

ratio

Dimensionality reduction ratio of PCA, default to 0.85.

Value

Subsample index.

Details

The uniform design is generated according to the method described in part 5 of Zhang's article. The uniformity measure uses mixed divergence.

References

Mei Zhang, Yongdao Zhou, Zheng Zhou & Aijun Zhang (2023) Model-Free Subsampling Method Based on Uniform Designs, IEEE Transactions on Knowledge and Data Engineering, 36:3, 1210-1220, https://ieeexplore.ieee.org/abstract/document/10192374.

Examples

data <- data_numeric_regression
X <- data[-which(names(data) == "y")]
DDS(X, n = 100, ratio = 0.85)
#>        [,1]
#>   [1,] 2238
#>   [2,] 2843
#>   [3,] 7010
#>   [4,] 8283
#>   [5,] 9340
#>   [6,] 1738
#>   [7,]  787
#>   [8,] 9555
#>   [9,] 1342
#>  [10,] 4120
#>  [11,] 9243
#>  [12,]  718
#>  [13,]  688
#>  [14,] 4792
#>  [15,] 3111
#>  [16,] 5176
#>  [17,] 6981
#>  [18,] 3831
#>  [19,] 4353
#>  [20,] 9894
#>  [21,] 3000
#>  [22,] 7966
#>  [23,]  311
#>  [24,] 8780
#>  [25,] 9324
#>  [26,] 1742
#>  [27,] 1390
#>  [28,] 8222
#>  [29,] 1827
#>  [30,] 7583
#>  [31,] 2335
#>  [32,] 9181
#>  [33,] 3109
#>  [34,] 2041
#>  [35,] 2395
#>  [36,] 6539
#>  [37,] 7812
#>  [38,] 5616
#>  [39,] 1753
#>  [40,] 5028
#>  [41,] 3147
#>  [42,]  460
#>  [43,] 7741
#>  [44,] 5427
#>  [45,]  670
#>  [46,] 3965
#>  [47,]  824
#>  [48,] 4025
#>  [49,]  406
#>  [50,] 9813
#>  [51,] 8686
#>  [52,] 6071
#>  [53,]   87
#>  [54,] 8553
#>  [55,]   85
#>  [56,] 2886
#>  [57,]  492
#>  [58,] 6377
#>  [59,] 3126
#>  [60,] 3064
#>  [61,]  423
#>  [62,] 8448
#>  [63,]  611
#>  [64,] 2565
#>  [65,] 6673
#>  [66,] 1971
#>  [67,] 9727
#>  [68,] 5396
#>  [69,] 3745
#>  [70,] 7576
#>  [71,] 4819
#>  [72,] 8385
#>  [73,] 4268
#>  [74,] 5835
#>  [75,] 6308
#>  [76,] 5660
#>  [77,] 7875
#>  [78,] 4666
#>  [79,] 4072
#>  [80,] 3578
#>  [81,] 2771
#>  [82,] 1801
#>  [83,] 4466
#>  [84,] 5332
#>  [85,] 4813
#>  [86,] 9383
#>  [87,] 1443
#>  [88,] 3293
#>  [89,] 4392
#>  [90,] 4804
#>  [91,] 2998
#>  [92,]    8
#>  [93,] 2750
#>  [94,] 6135
#>  [95,] 5014
#>  [96,]  657
#>  [97,] 5944
#>  [98,] 9737
#>  [99,] 2651
#> [100,] 2354