Orthogonal subsampling for big data linear regression (OSS, Rcpp-version by the package itself)

A subsampling method based on orthogonal array for linear model.

Usage

OSS(n, X)

Arguments

n

Subsample size.

X

A matrix or data frame.

Value

Subsample index.

References

Lin Wang, Jake Elmstedt, Weng Kee Wong & Hongquan Xu (2021) Orthogonal subsampling for big data linear regression, The Annals of Applied Statistics, 15(3), 1273-1290, https://projecteuclid.org/journals/annals-of-applied-statistics/volume-15/issue-3/Orthogonal-subsampling-for-big-data-linear-regression/10.1214/21-AOAS1462.short?tab=ArticleLink.

Examples

data_numeric_regression["y"] <- NULL
X <- as.matrix(data_numeric_regression)
OSS(100, X)
#>   [1] 8841 8961 1902 7512   48 9867 6547 9784 3392 3622 5780 6594 1890 1850 8335
#>  [16] 1254 6204 1257 4611 3831 4782 4919 1579 3404  718 7189 2060 4899  590 1800
#>  [31] 1195 3763 4033  607  714 8515 9026 4449 1921 6102 2550 8666 7403 1217 2633
#>  [46] 6481 5047  901 1738 9774 5242 1293 4981 5754 9770 7773 3075 9778 5650 1914
#>  [61] 7010 1605 6000 4995 6115  131 1594 7895 5217 2345 6992 9790 4461 4135 4737
#>  [76] 8989  129 8542 3067 3116 5266 3156 1365 8982 6922 5105 1366 5136 3532  253
#>  [91] 5693 8559 6242 2431  999 7480 5310  224 5662 1280